requirements.txt
pandas==0.24.1
matplotlib==2.2.2
numpy==1.14.5
psycopg2==2.7.6.1
python-dotenv==0.10.1
Scrapy==1.5.2
selenium==3.8.0
slackclient==1.1.0
ipython==5.8.0